+++ inspeqtor.go
@@ -101,11 +101,16 @@ func (i *Inspeqtor) Parse() error {
 	util.DebugDebug("Global config: %+v", config)
 	i.GlobalConfig = config
 
-	host, services, err := ParseInq(i.GlobalConfig, i.RootDir+"/conf.d")
+	host, err := ParseHost(i.GlobalConfig, i.RootDir+"/host.inq")
 	if err != nil {
 		return err
 	}
 	i.Host = host
+
+	services, err := ParseServices(i.GlobalConfig, i.RootDir+"/services.d")
+	if err != nil {
+		return err
+	}
 	i.Services = services
 
 	util.DebugDebug("Config: %+v", config)